Add hashtag links into the template
